Add files via upload
This is an R script for genotyping brachypodium hybridum. To use this script you will need to set the directory in the script to the location where the hapmap files are located. It is recommended that people familiar with R use this script and run it in sections. Running this script as is will generate lots of output files and images.