Add files via upload
This is a genotyping filter script for R, that I created in my PhD at ANU. If you use this script remember to change the directory. Also if you run this raw, it will create many output files and images. It would be best to be used by an experienced R user.