Add files via upload
This is a zipped file of genotype data for 528 odd B. distachyon samples and a few suspect samples. This is genotype data from my PhD at ANU. If you use this data you agree to publish after we have published this data set. Which we have not yet. Please contact us if you plan on using this data set.