Add files via upload
This is genotype data from my PhD at ANU. This works in tandem with a genotyping script I have posted in this repository. if you use this information understand that it not yet published and that if you use this data set you must contact us first before you publish and use this data set and get our approval here from the Borevitz lab at ANU. We are very pro open-source.