Add files via upload
This is genotype data for Brachypodium hybridum from my PhD at ANU. This data set is opensource, but we urge you to contact and make an agreement about how to use this data set before you publish it. Also you must wait until we publish this data set before you do and it is not published yet. This is a polyploid data set where non-unique reads have been removed. Each subgenome is attached. B. stacei  and B. distachyon like subgenomes are included.