Add files via upload
This is a hapmap text file from my PhD created 2016/12/26. This data set can not be published without first notifying and making an agreement with the Borevitz Lab at Australian National University and myself Jared Streich about publication use and date of publication submission data. This is posted for the purposes of my PhD submission for review.